bugfix: Freight is type of decimal, which translates to float (instead of int) in python

+++ order.py
@@ -5,6 +5,6 @@ from record_template import *
 class Order(Record):
     table_name = "Orders"
     fields = (("OrderID", int), ("CustomerID", str), ("EmployeeID", int), ("OrderDate", datetime),
-              ("RequiredDate", datetime), ("ShippedDate", datetime), ("ShipVia", int), ("Freight", int),
+              ("RequiredDate", datetime), ("ShippedDate", datetime), ("ShipVia", int), ("Freight", float),
               ("ShipName", str), ("ShipAddress", str), ("ShipCity", str), ("ShipRegion", str), ("ShipPostalCode", str),
               ("ShipCountry", str))
